Fixed beginSerial() to take baud rates above 19200; adjusted Serial Monitor Baud Rate menu items accordingly.

+++ targets/arduino/wiring.c
@@ -176,7 +176,7 @@ void analogWrite(int pin, int val)
 		digitalWrite(pin, HIGH);
 }
 
-void beginSerial(int baud)
+void beginSerial(long baud)
 {
 	uartInit();
 	uartSetBaudRate(baud);
